# solution/central-server/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Cinevault Central Server API Documentation",
    "description": "Documentation for the central server routes.\n\nContacts:\n- Giuseppe Falcone (giuseppe.falcone@edu.unito.it)\n- Gabriele Vettorello (gabriele.vettorello@edu.unito.it)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Central server"
    }
  ],
...
  "paths": {
...
    "/people/actor/{actorName}": {
      "get": {
        "summary": "Retrieve actor page with movies and awards",
        "description": "Displays movies featuring the actor along with any associated Oscar awards using the provided actor name.",
        "tags": [
          "People"
        ],
        "parameters": [
          {
            "name": "actorName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Leonardo DiCaprio"
            },
            "description": "Name of the actor"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content for the actor's page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing actor name",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPage"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPage"
                }
              }
            }
          }
        }
      }
    },
    "/people/crew/{crewName}": {
      "get": {
        "summary": "Retrieve crew member page with movies and awards",
        "description": "Displays movies the crew member worked on along with any associated Oscar awards using the provided crew name.",
        "tags": [
          "People"
        ],
        "parameters": [
          {
            "name": "crewName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Christopher Nolan"
            },
            "description": "Name of the crew member"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content for the crew member's page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing crew name",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPage"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorPage": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Internal Server Error"
          }
        }
      }
    }
  }
}